+++ cmps285HotelProject/HicsHotel/HicsHotel/Models/Customer.cs
@@ -23,7 +23,7 @@ namespace HicsHotel.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public int Phone { get; set; }
+        public double Phone { get; set; }
         public string Email { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
